refactoring: minor improvements
- Create new Pictures folder and move there writers,
  readers and PictureFactory
- Rename Pnm to PnmHandler
